Add support for `deserializedInterface` to AutoCodec
Added a field to `@AutoCodec` called `deserializedInterface`, which when set, creates a subclass of the annotated class to be used by all deserialized objects. This subclass will implement the given interface. Using this feature places certain restrictions on the annotated class (has to use a constructor as the instantiator, can't be final, can't be an inner class).

PiperOrigin-RevId: 809193075
Change-Id: I1545794f2bb32548601e44e0afb6934749fbf0fb