Don't add apple env if Xcode version isn't set
In the case no Xcode versions are installed, but something is requesting
this, like when you try to build an objc_library with only the CLT
installed, the `xcode_version` here is `None`. This fails with a
surprising crash in bazel when the `None` version is parsed. Now we just
don't set this, which cascades to later failures, but has slightly nicer
error messages.