Fix DefaultSyscallCache for normalizing filesystems
`DefaultSyscallCache` includes an optimization that answers a query for a file's type via a cached `readdir` result. This optimization didn't work correctly on macOS and Windows for various reasons: it hardcoded whether a file system on these OSes is case-insensitive, didn't reencode strings properly and also ignored Unicode normalization.

This is fixed by only making assumptions about the case mapping and normalization behavior of ASCII strings, which still allows this optimization to apply to the common case of a `BUILD` file existence check in a directory with only ASCII file names.

Along the way and since any OS now begins with an exact search in the list of `Dirent`s, migrate to `CompactSortedDirents` for reduced peak memory usage.

Closes #26842.

PiperOrigin-RevId: 810126203
Change-Id: I0dad35e3d45f6ea1185e2dd8dfdec8938805c407