Also use the Bazel string representation in the remote worker.
It's less error-prone to stick to a single convention throughout the codebase. (While the remote worker is an independent tool, it shares nontrivial parts of its codebase with the Bazel binary.)

Also remove the requirement that `--cas_path` must already exist; simply create it if not, like we already do for `--work_path`.

PiperOrigin-RevId: 810854163
Change-Id: Idd7e47c51e3a18c4bbd5b5ab29ad2de3acda3c30