Make sure `TreeArtifactValue.EMPTY` is deserialized into a `DeserializedSkyValue` object
Since `TreeArtifactValue.EMPTY` is marked as a `@SerializationConstant`, its deserialized version does not implement `DeserializedSkyValue` and fails the check in `SkyValueRetrieverUtils`.

This CL makes it so that `TreeArtifactValue` has its own `ObjectCodec` impl that explicitly deserializes empty instances into a special constant `EMPTY_DESERIALIZED`. This codec is a simple wrapper around the `@AutoCodec`-generated codec, which is specified to not `autoRegister`.

PiperOrigin-RevId: 811182550
Change-Id: Iaea6b1173ca13fc337aee416da28d64fa41bba27