Skycache: Do not check targets in metadata lookup requests
The Skycache metadata tables store the top level targets of the writer invocations. However there may be serialized targets below those top level targets within the frontier that would also get cache hits if used as top level target in a reader invocation.  The metadata info message would incorrectly say that no cache hits would be expected in that case. We were aware of this issue and it's solvable by adding all the serialized targets in the writer invocation to the metadata table, not just the top level ones.

However, it is harder to solve the scenario where the top level target in a reader invocation is a reverse dependency of a target cached in a writer invocation and this reader top level target doesn't appear anywhere in any writer invocations. In that case the reader can still get substantial cache hits but the metadata message would say "no cache hits expected".

The simplest thing we can do is to remove checking targets from the query and consider that it's enough to check the config checksum and CL. This will be true with a wide adoption of Skycache.

In follow up CLs I will rename the RPC and modify the message to be more clear about the expectaiton around cache hits.

RELNOTES:none
PiperOrigin-RevId: 811698965
Change-Id: I65215ad8ef6666674c1559a9e521d8387d5bdc19