Replace merging of `CcCompilationContext` from a builder to a static method
This works towards Starlarkification of CcCompilationContext. Static methods work better in Starlark than builder pattern.

Add createAndMerge method, which is a rewrite/inlininig of the Builder...build calls as used during merging (but not creation; creation is handled in next change).

Public method mergeCompilationContexts and private API createCcCompilationContextWithExtraHeaderTokens are migrated to use create or createAndMerge and avoid using Builder.

addIfNotNull and TransitiveSetHelper are moved one class higher, so they may be used.

The transformation is a no-op.

PiperOrigin-RevId: 811847335
Change-Id: Ic0deb4235d0966de829a376c810887b3d94fc247