Treat action results with missing outputs as cache misses
This allows builds to recover from a polluted remote cache that contains an action result with an exit code of 0 that hasn't produced the required output files.

The check added in a151116f7f671efea9a95f3e251561e53e535cca is moved closer to the lookup so that the cached result can be ignored instead of resulting in an error, with forced re-execution in the case of remote execution.