Retain ackQueue as request event messages
The queue of to-be-acknowledged events can grow indefinitely as a remote endpoint fails to deliver acknowledgements for event sequence numbers. This unbounded growth can leave references to resources associated with events in their rich form, particularly for file uploads (via futures), where Paths retain a reference to their FileSystem. Over the life of the build, this will prevent heap reclamation during garbage collection, and require heap retention relative to the number of actions' FileSystems, which will, in conjunction with Build without the Bytes remote execution, be RemoteActionFileSystems with unique maps of inputs.

This change reduces the footprint of the events ack queue strictly to the size of the message, which will at least by definition be less than the cost of retaining maps of all actions' inputs over the lifetime of the build, and ideally be a smaller footprint with less work to perform for ack sends of constructed messages on a requirement for retransmit.

Closes #22696.

PiperOrigin-RevId: 812848436
Change-Id: Iccd456591535c4319235d0cf47cd7ee575fef663