Fix forward https://github.com/bazelbuild/bazel/commit/1c07ab87448d06119ed44b76442b11aa055906a5
The culprit cl caused a slight difference in ordering of the headers, because they were added as a transitive set to Stable order instead of direct. Stable order is also known as topological.

I reviewed all fields part of "single" so that they are now merged the same way as previously. Some fields previously weren't direct/single. For those I added precondition check.

PiperOrigin-RevId: 813203229
Change-Id: I215919ee7812fd4bffe6be6cbaab36ceff565e6d