`ExpandedDirectoryApi` should not extend `FileApi`.
The original intention behind extending `FileApi` was to allow users to directly use it as though it was its underlying directory (e.g. passing it along as an input etc). But that was scrapped since we would need to special-case `ExpandedDirectory` when handling inputs during action creation (possibly expanding nestedsets) in order to resolve it to its underlying directory. As such, a `directory` API was added to `ExpandedDirectoryApi` to force users to explicitly reference the directory itself.

PiperOrigin-RevId: 813878757
Change-Id: I4d3ae4b6a1ac4e5b62da6e97935019e4fb1fb10f