`CompactPersistentActionCache`: don't use an error for rebuilds.
Log statements like the following that use `ERROR:` make it sound like the build
is going to fail, as I believe this is always what `ERROR` implies otherwise
(?):

```
ERROR: Error during action cache initialization: Failed action cache referential integrity check: validation record 722354 is too large compared to index size 722320. Data may be incomplete, potentially causing rebuilds
```

This seems inappropriate for something that is a bazel-internal problem (not the
user's fault), one that is automatically worked around, one that is not
reproducible, and one where the build ultimately succeeds. Downgrade it to a
warning.

--------------------------------------------------------------------------------

PiperOrigin-RevId: 817570402
Change-Id: Id276f82510e5ea0935933b9ff59010510f3804ab