Make external repo file checking actually useful
Bazel is checking all files under external repository directories for modifications compared to the previous in-memory state. Before this change, upon noticing modifications, it only updated its own in-memory state of the modified files' metadata, which is not correct: files in external repos are under exclusive control of repo rules and any modifications will result in incorrect incremental builds.

This change uses the diff check to show a warning whenever a file in an external repository is modified externally and trigger a refetch of the corresponding repository to recreate the expected state of the file.

RELNOTES: With `--experimental_check_external_repository_files` enabled (the default), Bazel will now refetch the respective repositories when it encounters external modifications. This is necessary to ensure correct incrementality. If you rely on external modifications to these repositories, either disable or the flag or use a supported mechanism such as `--override_repository`, `local_path_override` or `override_repo`.

Work towards #24694

Closes #27032.

PiperOrigin-RevId: 818659332
Change-Id: Ifc99377d48607b96808f45aeacb519181c17904a