Simplify NativePosixFiles.readdir().
By moving out followup stat calls and picking the most straightforward choice for the return type.

Note that virtually every modern filesystem is able to populate the `d_type` field on `struct dirent`, so followup stat calls are only needed when an entry is a symlink and follow semantics were requested (or for an unusual filesystem that does not provide file types).

Also future-proof the native implementation by returning types that Bazel currently doesn't care about.

PiperOrigin-RevId: 819257222
Change-Id: I33ffbb2083b07c92c7d926f05f4e7a8f2cc949c9