Add run command option to use current working directory
From the discussion in https://github.com/bazelbuild/bazel/issues/3325#issuecomment-2810070670, this adds a `--run_in_cwd` option to the `bazel run` command, which if enabled runs the target in the current working directory as opposed to running it in the runfiles directory.

The workaround by using `--run_under` to get a target to run in the current directory doesn't always work in every situation or is just unfortunate. This would improve the flow of running tools against source code for things like linting, formatting, running code generators, etc. which are present in `BUILD` files.

Closes #27290.

PiperOrigin-RevId: 820560189
Change-Id: Ib9cc17c9b05926a6cdb43d86c4ac22f9f3ee0500