[8.6.0] Add a `target_type` argument to `ctx.actions.symlink`.
This is necessary to create the right kind of filesystem object on Windows (junction for directories, symlink for files) when the target does not yet exist.

This argument is only allowed in conjunction with `target_path`. For `target_file`, I have a different plan (infer the type from the artifact) which will be implemented separately.

Fixes #26701.

Progress on #21747.

RELNOTES: `ctx.actions.symlink` now accepts a `target_type` argument.
PiperOrigin-RevId: 820670309
Change-Id: I2f29adfd074c404a0b15be369a97fcdfb84fbdad