[7.7.0] Invalidate subtrees when a directory is deleted (#27109)
`FileSystemValueCheckerInferringAncestors` claims that it is resilient
to diffs which only report the root of a deleted subtree, but this
wasn't true. Stale `FileStateValue`s and `DirectoryListingStateValue`s
could linger around after moving an entire directory tree out of the
workspace watched with `--watchfs`.

This is fixed by invalidating all FSVs and DLSVs under any deleted
directory. Since these values and their keys do not depend on their
parents, this requires a manual walk of the entire graph, making use of
a prefix-free sorted map to speed up lookups per key.

Fixes #26863
Fixes #26866

Closes #26920.

PiperOrigin-RevId: 812740160
Change-Id: Id55329b5ffd3cf3e109ec102e86b6fb46335531b

Commit
https://github.com/bazelbuild/bazel/commit/8df327a150a98bec1c77f71151a74406e1923320

---------

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>
Co-authored-by: Ian (Hee) Cha <heec@google.com>