Fix various errors in SevenZDecompressorTest
The test was using the same decompressed output directory that other tests were
using, a generic "out" directory. So test results were verifying/depending on
the output of those other tests.  Now the test uses /* outDirName= */
this.getClass().getSimpleName()

Fixing that revealed that the expected decompressed file was named wrong:

wrong: root_folder/another_folder/regular_file
right: root_folder/another_folder/regularFile

Instead of using `archiveDescriptor.assertOutputFiles`, which relied on the
output of other tests, we now inline the exact assertions we need.

Previously, the assertions would check that a file's `exists()` method was
true. This didn't produce the most friendly test output.  For example:

> value of: exists()
> expected to be false

Now we get a directory listing of entries and check for the file in that list:

> expected to contain: myfile
> but was            : [regularFile, renamedFile]

For the unicode test, we now make use of `StringEncoding::internalToUnicode`
when reading directory entries and `StringEncoding::unicodeToInternal` when
checking for a files existence. This is because meteorcloudy@ reported that
internal tests failed, which I am making an educated guess that it has
something to do with the string encoding.