Fix hermeticity of test fs
Each test was using the same output directory for the decompressed files (was this.getClass().getSimpleName()).
This would cause problems as the decompression of files in one test would affect other tests.

Now each test creates their own TestArchiveDescriptor with the output directory being set to:

> this.getClass().getSimpleName() + "_" + name.getMethodName()

As an example, for `testDecompressWithRenamedFiles` the directory would be:

> SevenZDecompressorTest_testDecompressWithRenamedFiles