Move CcCompilationContext implementation to Starlark
This change moves the internal data representation and the logic for creating and merging `CcCompilationContext` from Java to Starlark. The Java `CcCompilationContext` now acts as a wrapper around a `StarlarkInfo` provider. The `create_compilation_context` and `merge_compilation_contexts` functions are now implemented in `cc_info.bzl`.

This increases the surface of CcInfo implementation in Starlark. What's kept native for now is HeaderInfo.

PiperOrigin-RevId: 821807849
Change-Id: I1c137ad4bd87a76360554a059a056a7031ac2944