Starlarkify `CppModuleMap`.
The `CppModuleMap` class is now backed by a `StarlarkInfo` struct. The creation of `CppModuleMap` instances and the "separate module map" functionality are moved entirely to Starlark. Java code now accesses the `file` and `name` attributes from the underlying `StarlarkInfo`. This removes the need for `CppModuleMap` to implement `CppModuleMapApi` and simplifies the Java-Starlark boundary.

PiperOrigin-RevId: 822071919
Change-Id: Icb554683f9a6497e63b55598cb7bf5acbdf6b503