Remove unicode file and test for SevenZDecompressor
See discussion in #27269 - there was some internal testing difference for
unicode files that we weren't able to figure out.  It works locally on my local
mac computer but did not work inside Google for some reason.  Since this is the
only decompression test that tried to test unicode files, we decided to drop
it.  The thinking is that this would also occur in the other decompression
tests if they also had a unicode file.  My own integration testing shows that
it does work.