[8.5.0] Treat action results with missing mandatory outputs as cache misses (#27380)
This allows builds to recover from a polluted remote cache that contains
an action result with an exit code of 0 that hasn't produced the
required output files.

The check added in unknown commit is moved closer to the lookup so that
the cached result can be ignored instead of resulting in an error, with
forced re-execution in the case of remote execution.

Closes #27105.

PiperOrigin-RevId: 821651398
Change-Id: I067c668598f9ee83e670b4636854185282112f86 (cherry picked from
commit 29d442bc02e4829dc88ba5f79a3d626d1ca8680e)

Closes #27363