Handle edge cases for 7-zip decompressions
Two edge cases for 7-zip decompressions are handled:

1) When the entry in the .7z archive has no name
2) When the entry has no modification date

# Edge case 1 (entry has no name):

Previously, we would construct an output file that pointed to a directory and
the extraction would fail with a cryptic error message about trying to write to
a directory:

> /example/path (Is a directory)

Now we provide an explicit error message telling the user that

> 7z archive contains unnamed entry

We don't gracefully handle this since there are many options/cases:

 - When there is a single unnamed entry, should we follow the command-line
   behavior and name it a default name (the archive name)?

 - If there are multiple unnamed entries, do we
   1) overwrite previous entries
   2) keep the first entry
   3) Do some naming scheme, (unnamed_file1, unnamed_file2)

# Edge case 2 (entry has no modification date):

Previously, this would throw UnsupportedOperationException("The entry
doesn't have this timestamp"). Now, we don't set a modification date if there
isn't one.