Enable more HTTP decompression methods for remote cache
Bazel has support for compressed transfers from/to gRPC remote caches (https://github.com/bazelbuild/bazel/pull/14041). However, this is not fully supported for HTTP protocol. This feature would allow to significantly decrease the network traffic for remote cache. The related issue is #23463.

In general, the `HTTPCacheClient` class can decompress the incoming HTTP content because it uses internally the `HttpContentDecompressor` from Netty library. But current implementation allows only "gzip" as acceptable encoding.
The version of Netty library used in Bazel has been updated recently to 4.1.110. This version supports much more compression methods than before. The suggested change allows using the new compression options with HTTP remote cache. "Zstd" compression is of particular interest because it is used in popular `bazel-remote` software.

Note: The suggested change forces the `bazel-remote` to send HTTP content with "zstd" encoding. For compatibility we may consider applying the `--remote_cache_compression` option to enable this functionality. I leaved it as is for now because "gzip" was used unconditionally.

Closes #27271.

PiperOrigin-RevId: 823203959
Change-Id: I5d8bf11cf3cddad688084bf1e6f53b8782f92f12