Allow nonlocal tests with nobuild_runfile_manifests
Addresses https://github.com/bazelbuild/bazel/issues/23980

When testing a large number of targets against a build farm, runfile manifests can start to pile up pretty quickly on the client, and take up a large amount of disk space. The runfile manifests are also unnecessary when performing tests with the sandboxed or remote spawn strategies, only (to my understanding) local or standalone.

This change modifies the test wrappers to error early if the runfile directory can't be found, and instructs the user to build with runfile manifests, or change their execution strategy.

We still need to be able to actually get the runfile directory when the input manifest isn't present. To address this, we also change the behavior of getRunfilesDirectory to just return the runfilesTreeArtifact's path.