Simplify service registration and create a way for a service to supply options.
The service registry is not needed: the service component can simply pass in a list of instantiated services into BlazeRuntime.main(), and the lookup can be done by iterating the list (there will only be a few of them and the iteration is done at startup, so performance isn't a concern). The design mirrors the way BlazeModules work.

Also introduce OptionsSupplier, a super-interface of BlazeModule and BlazeService containing only the methods related to options. This provides a uniform way to collect them from both components.

PiperOrigin-RevId: 823940987
Change-Id: Ifc414b211d4afc79737941987ec1b6e315772175