Refactor Decompressor documentation and code
When new decompressors are added, the documentation and code has to change in
several different locations.  This refactoring removes some redundant code
changes by moving the mapping of file extensions to a more proper data
structure. This allows some of the redundant code to be written more concisely.
Following that, any documentation changes that are needed are also revamped to
reduce duplication.  Tests are added that will catch if documentation is NOT
update.