Add type param to repository_ctx.extract()
Before, you had to rename or symlink a file to the correct compressed/archived
file extension in order for it to be decompressed properly. Now you can specify
the type of the file explicitly, without renaming/symlinking.

In other words, before, this wouldn't work, because the file type could not be
determined:

  repository_ctx.extract("my_archive")

Now you can set a type (the archive is zipped):

  repository_ctx.extract("my_archive", type="zip")

Fixes: #26254