Upgrade rules_cc to 0.2.12 and adapt Bazel to use external providers.
This change updates the `rules_cc` dependency to version 0.2.12. To support the migration of `CcInfo` and `DebugPackageInfo` to the external `@rules_cc` module, Bazel's Java code now checks for these providers from both the builtins and the `@rules_cc` repository. Test files are updated to include the necessary load statements from `@rules_cc` and to use helper methods that can retrieve the providers from either location.

PiperOrigin-RevId: 824525797
Change-Id: Icb881657dde35549fa88d5fbfdc0058ad46363af