[8.5.0] Replace RxJava with virtual threads for `--remote_cache_async` (#27397)
All RxJava usage in `RemoteExecutionService` and `UploadManifest` is
replaced with regular `Future`s and virtual threads that block on them.
By closing the `ExecutorService` that runs these virtual threads in
`RemoteExecutionService#shutdown`, all async uploads can be awaited
without the need for additional tracking in the form of a `Phaser`.

This hopefully resolves the spurious hangs reported in #25232 and even
if not would provide new insights in the form of stack traces of blocked
virtual threads.

Along the way, this resolves the issue that upload counts reported on
cancellation could be incorrect and often even negative due to double
counting. This results in a minor change in behavior: Uploads are only
included in the count when they start, not already when the
corresponding `FindMissingBlobs` is sent. Present digests are reported
as being uploaded.

Speculatively fixes #25232

Closes #27179.

PiperOrigin-RevId: 822992095
Change-Id: Ieb82ab86739cebe25fd7e90a2bc70897fe53b817 
(cherry picked from commit 76dfcc5fe9599d381f69943d3fa8fc4c064035ac)

Closes #27248