[8.5.0] Add --incompatible_eagerly_resolve_select_keys (#27429)
Cherry-pick of
https://github.com/bazelbuild/bazel/commit/054536e26a19ac4ef80360fc218c69ad1124a35e:

This avoids a common gotcha in which a `select` in a macro with repo
names in string keys will generally not work correctly when loaded in
the context of a different Bazel (Bzlmod) module that uses a different
set of apparent names.

In the rare case where resolution relative to the loading `BUILD` file
is necessary (e.g. in Skylib's `selects.config_setting_group` or other
macros that generate `config_setting`s and `select`s referring to them),
the keys can be preprocessed with `native.package_relative_label`.

RELNOTES: With the new `--incompatible_eagerly_resolve_select_keys`
flag, the label string keys of `select` dicts in `.bzl` files are
resolved relative to the containing file instead of relative to the
BUILD file that ends up using the `select`. Use
`native.package_relative_label` if this is not desired.

For https://github.com/bazelbuild/bazel/pull/26216.
Closes
https://github.com/bazelbuild/bazel/issues/26223#issuecomment-3444916396

---------

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>