[8.8.0] Add variable type annotation syntax and resolver behavior
The new syntax is only usable with `--experimental_starlark_type_syntax`. When `--experimental_starlark_type_checking` is enabled, the type expressions in the annotations must represent valid types.

The two pieces of syntax are annotations in simple assignment statements:

```
x : T = ...
```

and annotations as lone statements, a new `VarStatement` node:

```
x : T
```

Annotations may not be used in assignments with anything besides a single identifier on the LHS; or in other binding uses such as the targets of `for` statements or comprehension clauses.

A `VarStatement` is considered to be a binding occurrence of a symbol; it ensures the symbol is in the current block even if there are no other assignments to it. For example, it makes the following trigger a runtime failure due to there being no assignment to the local variable:

```
x = 1
def f():
    # error regardless of which order the below two lines are in
    print(x)  # statically refers to local var x
    x : int   # shadows the outer x
f()
```

Only the first binding occurrence of a symbol may contain a type annotation. This means that if there is a `VarStatement` for a symbol, it is the only such `VarStatement`, and it appears prior to any assignment for that symbol, and no assignment statement for that symbol contains a type annotation. Similarly, if there is no `VarStatement`, only the first assignment to the symbol may have an annotation.

Because global variables may not be reassigned -- i.e. may not have more than one binding occurrence -- exported symbols generally cannot have `VarStatements`. An edge case is that it's possible for a global symbol to have *only* a `VarStatement` and no actual assignments, but such a variable cannot be initialized and therefore won't exist in the resulting `Module` object at runtime.

A `def` statement is considered to type-annotate the function symbol it introduces if its signature contains any parameter or return type annotations. This means that in the (contrived) case of redefinitions of local functions, only the first `def` may have type annotations.

`VarStatement`s accept block and trailing doc comments in the same manner as assignment statements.

Work toward #27371.

Parser.java:
- Move doc comment reading code for `AssignmentStatement` into a helper `maybeParseTrailingDocComment`, so we can reuse it for `VarStatement`.

Resolver.java:
- Invert the boolean return value of `bind()`, to better match the expectation from Java collection APIs like `Set#add`.
- Add a `hasType` param to `bind()` that causes it to signal an error if this isn't the first binding occurrence of the symbol in the block.
- Eliminate the convenience overload of `bind()`, it added more confusion than it was worth. Ditto for `createBindingsForLHS()`.
- During the visit() traversal, visit `VarStatement`'s target, even though it doesn't really do anything since it's just an identifier.

NodeVisitor.java:
- Add TODOs for visiting `VarStatement`, as well as type annotations in other pre-existing nodes.

ParserTest.java:
- added `parseStatementError()` helper
- made these helpers' return values include location info so they're available for assertions

PiperOrigin-RevId: 825198675
Change-Id: Id96148edd3c327ad66fa14e78451f5ec46331314