Add support for %bazel.version*% variables in bazelrc imports
Follows the proposal:
https://github.com/bazelbuild/bazel/issues/24043#issuecomment-3352747391 by
mattyclarkson@ to add support for the following variables when importing
additional bazelrc files.  Variables are followed by their values for the
hypothetical build label 9.4.2-pre.20251022.1:

- %bazel.version% - 9.4.2-pre.20251022.1
- %bazel.version.major% - 9
- %bazel.version.minor% - 4
- %bazel.version.patch% - 2
- %bazel.version.prerelease% - pre.20251022.1
- %bazel.version.buildmetadata% - (empty)
- %bazel.version.major.minor% - 9.4
- %bazel.version.major.minor.patch% - 9.4.2

Eg. If your bazel version was 8.4.2-pre.20251022.1 and your .bazelrc had the
following:

> try-import %bazel.version%.bazelrc
> import %bazel.version.major%.%bazel.version.minor%-%bazel.version.patch%.bazelrc
> try-import %bazel.version.major%.bazelrc

It would be evaluated to:

> try-import 8.4.2.bazelrc
> import 8.4-2.bazelrc
> try-import 8.bazelrc

# Implementation details:

## Build label extraction:

Before: The build label was extracted out only when calling `bazel --version`
and when running in client server mode.

After: Now we always extract the the build label out early in the start of main

## Piping of build label:

The build label is stored in the OptionProcessor so that it can be passed to
`RcFile` when parsing occurs.

## Parsing and mapping of variables:

The build label is parsed during `RcFile::ParseFile` into its constituent parts
(major, minor, patch).  This occurs EACH TIME an import statement is found in a
.rc file. This is wasteful, but the logic is simple enough that it shouldn't
matter. Interpolation of the build label happens first, followed by
`%workspace%`.

Fixes: #24043