Add Ellipsis token and node
The `Ellipsis` (`...`) token is only meaningful to type expressions, and is not allowed in ordinary value expressions. When `--experimental_starlark_type_syntax` is on and `--experimental_starlark_type_checking` is off, the parser will tolerate `...` in type expressions. The parser and evaluator never tolerate `...` in value expressions regardless of flags.

The only potential/intended use for this token is to help represent type expressions borrowed from Python syntax: `Callable[..., R]` and `Tuple[..., R]`. We do not intend to use it as a singleton value at runtime (e.g. the way it is sometimes seen in slice expressions in Python/numpy).

- Add lexer token kind and syntax node. Trivially update `NodeVisitor` and `NodePrinter`.
- Update lexer. `ELLIPSIS` is consumed greedily in preference to `DOT`. Floats will greedily consume a trailing dot rather than make it available to either `ELLIPSIS` or `DOT`. This seems to match Python behavior from some quick testing though I didn't verify it exhaustively.
- Update parser to statefully track whether we're in a type expression (regardless of whether that type expression is parsed into type-specific AST nodes or into an arbitrary Expression tree).

Work toward #27371.

PiperOrigin-RevId: 825573507
Change-Id: Iff59f989c3b86d531f9bf3b7244f81bbee8a39d6