[8.5.0] Pass target type when creating symlinks in `SymlinkAction`.
On Windows, we need to know whether a symlink points to a file or a directory in order to create it properly (and we cannot rely on the target's existence when building without the bytes).

Unfortunately, it's very difficult to write an integration test for this, as the creation of the symlink and the target artifact is inherently racy in a real build. The best we can do is add unit tests for SymlinkAction.

Fixes #21747.

PiperOrigin-RevId: 826001349
Change-Id: I75deba3872e924f2b4c20d0e693c582ef2403140
(cherry picked from commit ec564c2a79c082622a76f2e528ba1ad63de0a2f7)