Increase the sleep time to 7 seconds in non_daemon_exit_test.sh
This test is always going to suffer the risk of being flaky, but a 5 second
sleep in the test is cutting it too tight (the daemon thread being tested
runs for at least 5 seconds).

Running this test 200 times before this change leads to ~4 failures.
Running it after this change more leads to 0 failures.

PiperOrigin-RevId: 826478833
Change-Id: I2f52e6e6dfea49bbf7e1d3065dc64644e75b1e17