Gracefully handle acion outputs when collecting lost inputs
We previously tried to ensure that `BulkTransferException#getLostArtifacts` was never called with any exceptions mentioning non-inputs, but that still resulted in crashes. As the retrier logic in the remote spawn runner makes this error-prone and it is very unlikely that the `InputMetadataProvider` is actually missing metadata for an input, this change simplifies the logic by treating every `ActionInput` missing from it as an output and raising an error (possibly resulting in a local fallback or retry) instead of crashing.

Speculatively fixes #27229