Remove the fallback logic to make the source file readable in `FileSystemUtils.moveFile`.
This logic was originally  added for the benefit of `SandboxHelpers.moveOutputs` (which must handle the situation where a sandboxed action produces unreadable outputs) but the latter has since grown its own custom logic because of tight performance constraints. Other callsites move files that are under our control and don't need this fallback.

PiperOrigin-RevId: 827494716
Change-Id: I78e1c4e78cf786897a127df2869a687a6b7abf21